==> 1_processing/00_upload_wandb_artifacts.ipynb <==
# Preprocessing

import pandas as pd
import numpy as np
import os
import yaml
from tqdm.notebook import trange, tqdm

node = !hostname
if "sc" in node[0]:
    base_path = "/sc-projects/sc-proj-ukb-cvd"
else: base_path = "/data/analysis/ag-reils/ag-reils-shared/cardioRS"
print(base_path)

dataset_name = "211110_anewbeginning"
mapping_path = f"{base_path}/data/mapping"
dataset_path = f"{base_path}/data/2_datasets_pre/{dataset_name}"

from pathlib import Path
Path(dataset_path).mkdir(parents=True, exist_ok=True)

vocab_dir = f"{mapping_path}/athena"
concept = pd.read_csv(f"{vocab_dir}/CONCEPT.csv", sep='\t')

sct_icd = pd.read_csv(f"{mapping_path}/umls/der2_iisssccRefset_ExtendedMapFull_US1000124_20220301.txt", sep="\t")

sct_icd

relationship = pd.read_csv(f"{vocab_dir}/RELATIONSHIP.csv", sep='\t')

vocabulary = pd.read_csv(f"{vocab_dir}/VOCABULARY.csv", sep='\t')

# links omop concept_ids
concept_relationship = pd.read_csv(f"{vocab_dir}/CONCEPT_RELATIONSHIP.csv", sep='\t')

## Prepare Records TTE

import wandb
import pathlib

# project="recordgraphs"
project="retina"
entity="cardiors"
artifact_date = "220812"

def get_path_from_wandb(reference: str):
    path = pathlib.Path(reference.split("file://")[1])
    assert path.exists()
    return path

def read_artifact(run, artifact_name, type_name):
    
    artifact = run.use_artifact(f'cardiors/recordgraphs/{artifact_name}:latest', type=type_name)
    file_path = get_path_from_wandb(artifact.manifest.entries[artifact_name].ref)
    print(file_path)

    return file_path

def add_artifact(run, artifact_name, type_name, description, artifact_path, df):
   
    artifact = wandb.Artifact(artifact_name, type=type_name, 
                              description=description,
                             metadata = {"1_shape": f"{len(df)}x{len(df.columns)}",
                                         "2_cols": str(df.columns.to_list())})
    artifact.add_reference(f"""file://{artifact_path}""", artifact_name, checksum=True)
    run.log_artifact(artifact)

run = wandb.init(project=project, job_type="log_artifact", entity=entity, tags=["artifact"])

phecode_defs_path = "/sc-projects/sc-proj-ukb-cvd/data/mapping/phecodes/phecode_strings_V2.csv"

artifact_name = "phecode_definitions"
type_name = "prepare_phecodes"
description = "Definition of phecodes"
artifact_path = f"{dataset_path}/artifacts/{artifact_name}_{artifact_date}.feather"

phecode_defs = pd.read_csv(phecode_defs_path, dtype={"phecode": str}).sort_values("phecode").reset_index()
phecode_defs.to_feather(artifact_path)

add_artifact(run, artifact_name, type_name, description, artifact_path, phecode_defs)

phecode_mapping_path = "/sc-projects/sc-proj-ukb-cvd/data/mapping/phecodes/ICD10_to_phecode_V2.csv"

artifact_name = "icd10_phecode_mapping"
type_name = "prepare_phecodes"
description = "Mapping of ICD10 to phecodes"
artifact_path = f"{dataset_path}/artifacts/{artifact_name}_{artifact_date}.feather"

phecode_mapping = pd.read_csv(phecode_mapping_path, dtype={"phecode": str}).rename(columns={"icd10": "ICD10"})\
    .assign(ICD10 = lambda x: x.ICD10.str.replace(".", "", regex=False)).reset_index()
phecode_mapping.to_feather(artifact_path)

add_artifact(run, artifact_name, type_name, description, artifact_path, phecode_mapping)

records_raw = pd.read_feather(read_artifact(run, "complete_records_extended", "prepare_records"))\
    .query("concept_id==concept_id")\
    .assign(concept_id = lambda x: x.concept_id.astype(int))
records_raw.origin.unique()

1+1

### Explicitly encode death record

records_death = records_raw.query('origin=="death_records"').copy()
#records_death.loc[records_death.query('origin=="death_records"').index, 'code'] = 4306655
records_death.loc[records_death.query('origin=="death_records"').index, 'concept_id'] = 4306655
records_death = records_death.drop_duplicates(subset=['eid', 'concept_id'], keep='first')
records_raw_w_death = pd.concat([records_raw, records_death], axis=0)

records_raw_w_death

### Prepare OMOP

records_omop = records_raw_w_death.assign(concept_code=lambda x: x.concept_id, vocabulary="OMOP")
records_omop["vocabulary"] = records_omop["vocabulary"].astype("category")

records_omop.head()

### Prepare OMOP -> ICD10

omop_counts = records_raw[["concept_id"]].value_counts()

omop_concepts = omop_counts.to_frame().reset_index()[["concept_id"]]
omop_concepts_md = omop_concepts.merge(concept[["concept_id", "concept_code", "concept_name", "vocabulary_id", "domain_id", "concept_class_id"]], on="concept_id", how="left").merge(omop_counts.to_frame().reset_index(), on="concept_id", how="left")

omop_concepts_md

omop_counts

sct_icd

def map_omop_icd10():
    concepts_in = omop_concepts.concept_id.to_list()
    concepts_out = concept.query("vocabulary_id=='ICD10'").concept_id.to_list()
    mapping = concept_relationship.query("concept_id_1==@concepts_in&concept_id_2==@concepts_out")\
        .merge(concept, left_on="concept_id_2", right_on="concept_id")
    mapping_clean = mapping[["concept_id_1", "concept_code"]]\
        .rename(columns={"concept_id_1":"concept_id"}).assign(vocabulary="ICD10")\
        .assign(concept_code = lambda x: x.concept_code.str.replace(".", "", regex=False))
    return mapping_clean

mapping_omop_icd10 = map_omop_icd10()

mapping_omop_icd10#.query("relationship_id=='Value mapped from'")

### Prepare ICD10 -> Phecodes

icd10_phecode_mapping = phecode_mapping.rename(columns={"phecode": "concept_code"})[["ICD10", "concept_code"]].assign(vocabulary="phecode")

mapping_icd10_phecode = mapping_omop_icd10.rename(columns={"concept_code": "ICD10"})[["concept_id", "ICD10"]]\
    .merge(icd10_phecode_mapping, on="ICD10")[["concept_id", "concept_code", "vocabulary"]]

mappings_omop_icd10_phecode = pd.concat([mapping_omop_icd10, mapping_icd10_phecode], axis=0).reset_index(drop=True)
mappings_omop_icd10_phecode["concept_code"] = mappings_omop_icd10_phecode["concept_code"].astype("category")
mappings_omop_icd10_phecode["vocabulary"] = mappings_omop_icd10_phecode["vocabulary"].astype("category")
mappings_omop_icd10_phecode = mappings_omop_icd10_phecode.set_index("concept_id")

mappings_omop_icd10_phecode

mappings_omop_icd10_phecode.dtypes

### Join ICD10 and phecode mappings

records_icd10_phecode = records_raw_w_death.set_index("concept_id")\
    .merge(mappings_omop_icd10_phecode, left_index=True, right_index=True)

records_icd10_phecode = records_icd10_phecode.reset_index()\
    .query("concept_code==concept_code")\
    .drop_duplicates(subset=["eid", "date", "concept_code", "vocabulary", "origin"], keep="first")

### Merge all mapped records

records_all_vocabs = pd.concat([records_omop, records_icd10_phecode], axis=0).reset_index(drop=True)

records_all_vocabs.info()

records_mapped_unique = records_all_vocabs.assign(concept_id=lambda x: x.vocabulary +"_" +x.concept_code.astype(str)).reset_index(drop=True).drop(columns=["concept_code"])

# make everything faster
for col in tqdm(records_mapped_unique.columns): 
    if "date" in col: records_mapped_unique[col] = records_mapped_unique[col].astype("datetime64[D]")

# make everything faster
for col in tqdm(["domain_id", "code", "concept_id", "origin", "vocabulary"]): 
    records_mapped_unique[col] = records_mapped_unique[col].astype("category")

records_mapped_unique = records_mapped_unique.drop_duplicates().reset_index(drop=True)

records_mapped_unique.shape

records_mapped_unique.info()

def add_artifact(run, artifact_name, type_name, description, artifact_path, df):
   
    artifact = wandb.Artifact(artifact_name, type=type_name, 
                              description=description,
                             metadata = {"1_shape": f"{len(df)}x{len(df.columns)}",
                                         "2_cols": str(df.columns.to_list())})
    artifact.add_reference(f"""file://{artifact_path}""", artifact_name, checksum=True)
    run.log_artifact(artifact)

# log in wandb
artifact_name = "final_records_omop"
type_name = "prepare_records"
description = "Cleaned first patient records from the data portal in long format"
artifact_path = f"{dataset_path}/artifacts/{artifact_name}_{artifact_date}.feather"
records_mapped_unique.to_feather(artifact_path)
add_artifact(run, artifact_name, type_name, description, artifact_path, records_mapped_unique)

# Upload baseline file

baseline_outcomes = '/sc-projects/sc-proj-ukb-cvd/results/projects/22_retina_phewas/data/baseline_outcomes_220627.feather'
bo_df = pd.read_feather(baseline_outcomes)


bo_df.head()

artifact_name = "baseline_outcomes"
type_name = "baseline_file"
description = "Clean outcomes from dataportal at time of recruitment"
artifact_path = f"{dataset_path}/artifacts/{artifact_name}_220627.feather"
bo_df.to_feather(artifact_path)
add_artifact(run, artifact_name, type_name, description, artifact_path, bo_df)

artifact = run.use_artifact('cardiors/retina/eids:v0', type='prepare_splits')
artifact_dir = artifact.download()
artifact_dir

def add_artifact(run, artifact_name, type_name, description, artifact_path, df):
   
    artifact = wandb.Artifact(artifact_name, type=type_name, 
                              description=description, metadata = {"1_shape": "manual upload"})
    artifact.add_reference(f"""file://{artifact_path}""", artifact_name, checksum=True)
    run.log_artifact(artifact)

import yaml
eids_path = '/sc-projects/sc-proj-ukb-cvd/data/2_datasets_pre/211110_anewbeginning/artifacts/eids_211209.yaml'
eid_dict = yaml.load(open(eids_path), Loader=yaml.CLoader)

run = wandb.init(project=project, job_type="log_artifact", entity=entity, tags=["artifact"])

artifact_name = "eids"
type_name = "prepare_splits"
description = "Dictionary containing the eids based on the assessment centres and train, valid and test splits"
# artifact_path = f"{dataset_path}/artifacts/{artifact_name}_211209.feather"
add_artifact(run, artifact_name, type_name, description, eids_path, eid_dict)

eid_dict = '/sc-projects/sc-proj-ukb-cvd/data/2_datasets_pre/211110_anewbeginning/artifacts/eids:v0'


# add covariate file

artifact = run.use_artifact('cardiors/RecordGraphs/baseline_covariates:latest', type='prepare_covariates')
artifact_dir = artifact.download()
artifact_dir

x = pd.read_feather('/sc-projects/sc-proj-ukb-cvd/data/2_datasets_pre/211110_anewbeginning/baseline_covariates_220503.feather')
#                     /sc-projects/sc-proj-ukb-cvd/data/2_datasets_pre/211110_anewbeginning/baseline_covariates_220323.feather')

def add_artifact(run, artifact_name, type_name, description, artifact_path, df):
   
    artifact = wandb.Artifact(artifact_name, type=type_name, 
                              description=description,
                             metadata = {"1_shape": f"{len(df)}x{len(df.columns)}",
                                         "2_cols": str(df.columns.to_list())})
    artifact.add_reference(f"""file://{artifact_path}""", artifact_name, checksum=True)
    run.log_artifact(artifact)

# log in wandb
artifact_name = "baseline_covariates"
type_name = "prepare_covariates"
description = "Dataframe of covariates at recruitment"
artifact_path = f"{dataset_path}/artifacts/{artifact_name}_220503.feather"
x.to_feather(artifact_path)
add_artifact(run, artifact_name, type_name, description, artifact_path, x)
