==> 1_processing/03_create_reference_data.ipynb <==
import pandas as pd
import numpy as np
import pathlib
from tqdm.auto import tqdm

import hydra
from omegaconf import DictConfig, OmegaConf

import torch
from pytorch_lightning import seed_everything

import ray

node = !hostname
if "sc" in node[0]:
    base_path = "/sc-projects/sc-proj-ukb-cvd"
else: 
    base_path = "/data/analysis/ag-reils/ag-reils-shared/cardioRS"
print(base_path)

project_label = "22_retina_phewas"
project_path = f"{base_path}/results/projects/{project_label}"
figure_path = f"{project_path}/figures"
output_path = f"{project_path}/data"

pathlib.Path(figure_path).mkdir(parents=True, exist_ok=True)
pathlib.Path(output_path).mkdir(parents=True, exist_ok=True)

output_path

## Get Data

# get stuff from wandb and create symlinks to current folder!!

from hydra import compose, initialize
from omegaconf import OmegaConf
hydra.core.global_hydra.GlobalHydra().clear()

initialize(config_path="../../ehrgraphs/config")
args = compose(config_name="config", overrides=["datamodule.partition=0", 
                                                "datamodule.use_top_n_phecodes=10000",
                                                "setup.use_data_artifact_if_available=False",
                                                "datamodule/covariates='no_covariates'",
                                                "datamodule.t0_mode=recruitment",
                                                
                                               ])
print(OmegaConf.to_yaml(args))

def extract_records_events_times(args):
    
    records_list = []
    outcomes_list = []
    
    # prepare extraction
    datamodule, _, _ = setup_training(args)
    
#     record_cols = datamodule.record_cols
    label_cols = list(datamodule.label_mapping.keys())
    
    for s in tqdm(["train", "valid", "test"]):
        eids = datamodule.eids[s]
        
        if s=="train":  dataset = datamodule.train_dataloader(shuffle=False, drop_last=False).dataset
        if s=="valid":  dataset = datamodule.val_dataloader().dataset
        if s=="test":  dataset = datamodule.test_dataloader().dataset

        # extract records
#         records_temp = pd.DataFrame.sparse.from_spmatrix(dataset.records, index=eids, columns=[f"{c}" for c in record_cols]).rename_axis("eid")
#         records_list.append(records_temp)

        # extract exclusion & events
        exclusions_df = dataset.exclusions
        events_df = dataset.labels_events

        times = dataset.labels_times
        censorings = dataset.censorings

        no_event_idxs = times == 0
        print(censorings.shape)
        print(times.shape)
        print(no_event_idxs.shape)
        print(censorings.values.repeat((1, times.shape[1])))
        
        times[no_event_idxs] = censorings.values.repeat((1, times.shape[1]))[no_event_idxs]
        
        1/0

        times_df = pd.DataFrame(data=times, index=eids, columns=[f"{c}_time" for c in label_cols]).rename_axis("eid")

        outcomes_temp = pd.concat([exclusions_df, events_df, times_df], axis=1)
        outcomes_list.append(outcomes_temp)
        
#     records_df = pd.concat(records_list, axis=0)
    outcomes_df = pd.concat(outcomes_list, axis=0)
        
    return outcomes_df

from retinalrisk.training import setup_training
seed_everything(0)

#args = compose(config_name="config", overrides=[f"datamodule.partition={partition}"])
records_df, outcomes_df = extract_records_events_times(args)

## Write Records

records_df.info()

for c in tqdm(records_df.columns):
    records_df[c] = records_df[c].astype(bool).sparse.to_dense()

records_df = records_df.sort_index()

records_df.info()

records_df.reset_index().to_feather(f"{output_path}/baseline_records_{artifact_date}.feather")

## Write Outcomes

for c in tqdm(outcomes_df.columns):
    if c.endswith("_prev") or c.endswith("_event"):
        outcomes_df[c] = outcomes_df[c].astype(bool).sparse.to_dense()
    if c.endswith("_time"):
        outcomes_df[c] = outcomes_df[c].astype(np.float32)

outcomes_df = outcomes_df.sort_index()

outcomes_df.info()

outcomes_df.reset_index().to_feather(f"{output_path}/baseline_outcomes_{artifact_date}.feather")

### Outcomes long

endpoints = sorted(outcomes_df.columns.str.replace("_prev|_event|_time", "", regex=True).unique().tolist())

outcomes_long = pd.DataFrame()

outcomes_df_list = []
cols = ["prev", "event", "time"]
for e in tqdm(endpoints):
    temp = outcomes_df[[f"{e}_{c}" for c in cols]].assign(endpoint = e)
    temp.columns = cols + ["endpoint"]
    outcomes_df_list.append(temp)

outcomes_long = pd.concat(outcomes_df_list,
                          axis=0)[["endpoint"] + cols].assign(endpoint = lambda x: x.endpoint.astype("category")).reset_index()

outcomes_long.info()

outcomes_long.to_feather(f"{output_path}/baseline_outcomes_long_{artifact_date}.feather")
